Update README.md
# ERR422-For-mcbe-INDEV-
I am currently working on a project to port ERROR 422 into Minecraft Bedrock Edition. (WORK USABILITY IN README FILE). When I first heard about ERROR422 Minecraft, I thought it was fake, And I decided to read an article about it, and it was real. I decided to play this cursed version. It was one of the most interesting stuff it could be. 
So I decided to port it into Minecraft Bedrock Edition, because of how interesting it was.
I am trying to release this project as soon as possible, so I can work on other stuff I find cool.
If you are reading this, try leaving any comments on this project.